Revert "Moved tests"
This reverts commit 7dc4b506018df67773f433aa09f24610dc1d648a.
